Idle the polling thread until it's time to wake up.
This significantly reduces the CPU load from polling kinesis. We were
looping poll! as fast as we could before, even if it was too early to
hit the endpoint again, and there is no data. Now we sleep the thread
until the poll times out, or the poll interval is ready.